sadasd

+++ modules/employee/scripts/controllers/EmployeeCtrl.js
@@ -8,87 +8,89 @@
 	
 	function controller($http, $scope, $alert, $state, $stateParams, $location, actionServices) {
 		console.log('EmployeeCtrl Loaded');
-		
+		//Model
 		$scope.model = 'employee';
-        $scope.Religions = [{"id": 1, "name": "Islam"},
-									{"id": 2, "name": "Katholik"},
-									{"id": 3, "name": "Protestan"},
-									{"id": 4, "name": "Hindu"},
-									{"id": 5, "name": "Buddha"}];
-		$scope.Marital = [{"name":"Maried"},{"name":"single"}];
-		$scope.WorkStatus = [{"name":"Tetap"},{"name":"Kontrak"},{"name":"Resign"}];
+		//List Fungsi
 		$scope.addOnClick = addOnClick;
 		$scope.updateOnClick = updateOnClick;
 		$scope.DeleteOnClick = DeleteOnClick;
 		$scope.religion ={};
 		$scope.marital ={};
 		$scope.workStatus ={};
+		$scope.province = {};
+		$scope.departement = {};
+		$scope.idJob = {};
+		//Load Data
+		$scope.Religions = [{"id": 1, "name": "Islam"},
+									{"id": 2, "name": "Katholik"},
+									{"id": 3, "name": "Protestan"},
+									{"id": 4, "name": "Hindu"},
+									{"id": 5, "name": "Buddha"}];
+		$scope.Marital = [{"name":"Maried"},{"name":"single"}];
+		$scope.WorkStatus = [{"name":"Tetap"},{"name":"Kontrak"},{"name":"Resign"}];
+		$scope.Province = [{"name":"Aceh"},{"name":"Sumatra Utara"},{"name":"Sumatra Barat"},
+							{"name":"Riau"},{"name":"Jambi"},{"name":"Kepulauan Riau"},
+							{"name":"Bangka Belitung"},{"name":"Bengkulu"},{"name":"Lampung"},
+							{"name":"DKI Jakarta"},{"name":"Banten"},{"name":"Jawa Barat"},
+							{"name":"Jawa Tengah"},{"name":"Jawa Timur"},{"name":"Daerah Istimewa Yogyakarta"},
+							{"name":"Bali"},{"name":"Nusa Tenggara Barat"},{"name":"Nusa Tenggara Barat"},
+							{"name":"Kalimantan Barat"},{"name":"Kalimantan Tengah"},{"name":"Kalimantan Selatan"},{"name":"Kalimantan Timur"},
+							{"name":"Kalimantan Utara"},{"name":"Sulawesi"}];
+		$http.get('/api/departement/').then(function(response) {
+                    //console.log('response', response.data.data);
+                    $scope.Departement = response.data.data;
+                });
+		$http.get('/api/jobtitle/').then(function(response) {
+                    //console.log('response', response.data.data);
+                    $scope.JobTitle = response.data.data;
+                });
+		function findDepartement() {
+			return $http.get('/api/departement/').then(function(response) {
+                    //console.log('response', response.data.data);
+                    $scope.Departement = response.data.data;
+                });
+		}
+		function findJobtitle() {
+			return $http.get('/api/jobtitle/').then(function(response) {
+                    //console.log('response', response.data.data);
+                    $scope.JobTitle = response.data.data;
+                });
+		}
 		function addOnClick() {
 			//var data = {};
 			var data = {
+				employeeNumber:$scope.employeeNumber,
 				idEmployee:$scope.idEmployee,
 				firstName:$scope.firstName,
 				lastName:$scope.lastName,
 				gender:$scope.gender,
 				birthDate:$scope.birthDate,
 				birthPlace:$scope.birthPlace,
-				religion:$scope.religion.selected.name,
-				marital:$scope.marital.selected.name,
-				citizenship:$scope.citizenship,
-				personalEmail:$scope.personalEmail,
-				companyEmail:$scope.companyEmail,
-				telephoneNumber:$scope.telephoneNumber,
-				mobileNumber:$scope.mobileNumber
-			}
-			console.log('coba data:',data);
-			if(!data.idEmployee) execCreate(data);
-			if(data.idEmployee) execUpdate(data);
-			/*$http.post('/api/' + $scope.model + '/?' , data).then(function(response){
-    	            console.log('response', response);
-    	            if(response.data.status == 'success'){
-                        if(done) return done(null, response.data.data);
-    	            }
-    	            if(response.data.status == 'error'){
-                        if(done) return done('error');
-    	            }
-    	        });
-			$location.path("/app/employee/list");*/
-			//$scope.create($scope.model,data);
-		}
-		function execCreate(data) {
-			// body...
-			console.log('execCreate loaded');
-			console.log('execCreate - data: ', data);
-			/*$http.post('/api/' + $scope.model + '/?' , data).then(function(response){
-				//vm.modalFormTitle = 'Edit';
-				$scope.dataEmployee = response.data;
-				alert('successfully created');
-				refresh();
-			}).error(function(response){
-				console.log('error - response: ', response);
-				alert(response.message);
-			});*/
-		}
-		function updateOnClick() {
-			//var data = {};
-			var data = {
-				firstName:$scope.firstName,
-				lastName:$scope.lastName,
-				gender:$scope.gender,
-				birthDate:$scope.birthDate,
-				birthPlace:$scope.birthPlace,
-				religion:$scope.religion,
-				marital:$scope.marital,
+				religion:$scope.religion.selected,
+				marital:$scope.marital.selected,
 				citizenship:$scope.citizenship,
 				personalEmail:$scope.personalEmail,
 				companyEmail:$scope.companyEmail,
 				telephoneNumber:$scope.telephoneNumber,
 				mobileNumber:$scope.mobileNumber,
-				workStatus:$scope.workStatus
+				workStatus:$scope.workStatus.selected,
+				motherName:$scope.motherName,
+				accountNumber:$scope.accountNumber,
+				province:$scope.province.selected,
+				city:$scope.city,
+				streetName:$scope.streetName,
+				zipcode:$scope.zipcode,
+				idJob:$scope.idJob.selected,
+				idDepartement:$scope.departement.selected,
+				assignDate:$scope.assignDate,
+				assignStart:$scope.assignStart,
+				assignFinish:$scope.assignFinish,
+				resignDate:$scope.resignDate,
+				detail:$scope.detail
 			}
 			console.log('coba data:',data);
-			console.log('url :','/api/' + $scope.model + '/?'+$scope.idEmployee);
-			/*$http.put('/api/' + $scope.model + '/?'+$scope.idEmployee , data).then(function(response){
+			
+			$http.post('/api/' + $scope.model + '/?' , data).then(function(response){
     	            console.log('response', response);
     	            if(response.data.status == 'success'){
                         if(done) return done(null, response.data.data);
@@ -96,28 +98,61 @@
     	            if(response.data.status == 'error'){
                         if(done) return done('error');
     	            }
-    	        });*/
-			//$location.path("/app/employee/list");
+    	        });
+			refresh();
+			$location.path("/app/employee/list");
+			
+		}
+		function updateOnClick() {
+			var data = {};
+			var coba = $scope.data;
+			console.log('coba data:',coba);
+			console.log('url :','/api/' + $scope.model + '/idEmployee?'+ $scope.data.idEmployee);
+			
+			$http.post('/api/' + $scope.model + '/?idEmployee='+$scope.data.idEmployee , coba).then(function(response){
+    	            console.log('response', response);
+    	            if(response.data.status == 'success'){
+                        if(done) return done(null, response.data.data);
+    	            }
+    	            if(response.data.status == 'error'){
+                        if(done) return done('error');
+    	            }
+    	            refresh();
+    	        });
+			$location.path("/app/employee/list"); 
+			
 			//$scope.create($scope.model,data);
 		}
 		function DeleteOnClick(id){
 			var model = $scope.model;
 			console.log('alamat :','/api/' + $scope.model + '?idEmployee='+id);
-			$scope.delete(model,id);
-			/*
-			$http.delete('/api/' + $scope.model + '?idEmployee='+id).then(function(response){
+			//actionServices.delete(model,id);
+			if (confirm('Are you sure you want to delete this?')) {
+				$http.delete('/api/' + $scope.model + '?idEmployee='+id).then(function(response){
     	            console.log('Delete response', response);
                     $scope.data = response.data.data;
     	        });
-			$location.path("/app/employee/list");
-			*/
+				refresh();
+				$location.path("/app/employee/list");
+    		}
+    		
+			
+		}
+		function refresh(){
+			return $http.get('/api/' + $scope.model + '/').then(function(response) {
+                    console.log('response', response.data.data);
+                    $scope.data = response.data.data;
+                });
 		}
 		$scope.view = {
 			list: function(){
+				console.log('workStatus',$scope.WorkStatus);
 				$scope.pageTitle = 'Employees List';
 				$scope.detailUrl = '#/' + $scope.model + '/edit';
 				$scope.viewUrl = '#/app/' + $scope.model + '/view';
-                                
+                findDepartement();
+				findJobtitle();
+
 				$scope.dataName = $scope.model + 'records';
 				$scope.loading = {};
 				$scope.loadingdata = true;
@@ -130,22 +165,23 @@
 			      {id: 6, firstName: 'Antonius', lastName: 'Simamora', gender: 1, genderLONG: 'Laki-Laki', birthDate: new Date('1981-12-24'), phone1: '0815-7431-4082', email1: 'anton@aksimaya.co.id'} 
 			  	];      
 			  	//data = $scope.rowCollections; 
-
+			  	
 			  	return $http.get('/api/' + $scope.model + '/').then(function(response) {
                     console.log('response', response.data.data);
                     $scope.data = response.data.data;
                 });
-
 			  	console.log('list d:',$scope.data);
-			  	console.log('list r:',$scope.rowCollections);         
+			  	//console.log('list r:',$scope.rowCollections);         
 			},
 			add: function(){
+				findDepartement();
+				findJobtitle();
 				$scope.pageTitle = 'Add New Employee';
-				
+				console.log('workStatus',$scope.WorkStatus);
 				$scope.saveBtnTxt = 'Save';
 				$scope.cancelBtnTxt = 'Cancel';
 				$scope.cancelBtnHref = '#/' + $scope.model + '/list';
-				                
+				//$scope.assignDate = {};
 				$scope.dataName = $scope.model;
 				$scope.data[$scope.dataName] = {};
 				//----- START: get data & init default select Static Option / enums -----//
@@ -156,6 +192,8 @@
 				$scope.initFormAdd();
 			},
 			view: function(id){
+				findDepartement();
+				findJobtitle();
 				$scope.pageTitle = 'View Employee Data';
 				
 				$scope.editBtnTxt = 'Edit';
@@ -168,14 +206,15 @@
 				console.log(id);
 				//
 				$http.get('/api/' + $scope.model + '?idEmployee='+id).then(function(response){
-    	            console.log('View response', response.data.data);
-                    $scope.data = response.data.data;
+    	            console.log('View response', response.data.data[0]);
+                    $scope.data = response.data.data[0];
     	        });
 				//$scope.findOneAndInitForm(id, {});             
 			},
 			edit: function(id){
 				$scope.pageTitle = 'Edit Employee Data';
-				
+				findDepartement();
+				findJobtitle();
 				$scope.updateBtnTxt = 'Update';
 				$scope.cancelBtnTxt = 'Cancel';
 				$scope.cancelBtnHref = '#/' + $scope.model + '/list';
@@ -183,9 +222,16 @@
 				$scope.mode = 'edit';
 				$scope.itemId = id;
 				$http.get('/api/' + $scope.model + '?idEmployee='+id).then(function(response){
-    	            console.log('Edit response', response.data.data);
-                    $scope.data = response.data.data;
-                    console.log('Edit response', $scope.data);
+    	            console.log('Edit response', response.data.data[0]);
+                    $scope.data = response.data.data[0];
+                    $scope.data.religion = {id: null, name: response.data.data[0].religion};
+                    $scope.data.marital = {name: response.data.data[0].marital};
+                    $scope.data.province = {name: response.data.data[0].province};
+                    $scope.data.workStatus = {name: response.data.data[0].workStatus};
+                    $scope.data.departement = {idDepartement:response.data.data[0].idDepartement ,departementName: response.data.data[0].departementName};
+                    $scope.data.idJob = {jobId:response.data.data[0].idJob,jobTitle: response.data.data[0].jobTitle};
+                    //$scope.data.idJob = {jobId: response.data.data[0].idJob,jobTitle: $scope.Departement.};
+                   	//console.log('religion',$scope.data.religion);
     	        });
 				//$scope.findOneAndInitForm(id, {});
 				//console.log($scope);
